+++ apps/site/assets/ts/schedule/__tests__/ScheduleModalTest.tsx
@@ -2,7 +2,6 @@ import React from "react";
 import renderer, { act } from "react-test-renderer";
 import { EnhancedRoute, RouteType } from "../../__v3api";
 import ScheduleModalContent, {
-  reducer,
   fetchData
 } from "../components/schedule-finder/ScheduleModalContent";
 import { SimpleStop } from "../components/__schedule";
@@ -183,33 +182,6 @@ describe("ScheduleModal", () => {
           type: "FETCH_ERROR"
         });
       });
-    });
-  });
-
-  describe("reducer", () => {
-    it("handles a fetch complete action", () => {
-      const newState = reducer(
-        { error: false, isLoading: true, data: null },
-        { type: "FETCH_COMPLETE", payload: [] }
-      );
-      expect(newState).toEqual({ data: [], isLoading: false, error: false });
-    });
-
-    it("handles a fetch error action", () => {
-      const newState = reducer(
-        { error: false, isLoading: true, data: null },
-        { type: "FETCH_ERROR" }
-      );
-      expect(newState).toEqual({ data: null, isLoading: false, error: true });
-    });
-
-    it("handles an unknown action", () => {
-      const newState = reducer(
-        { error: false, isLoading: true, data: null },
-        // @ts-ignore
-        { type: "UNKNOWN" }
-      );
-      expect(newState).toEqual({ data: null, isLoading: true, error: false });
     });
   });
 });
